Add interfaces and store functor
Signatures in interfaces.ml file to specify interfaces for
expressions and stores for use in functors etc.

The store functor takes an expression module and gives a store module
for that expression implementation. Required the adding of a couple
of extra functions such as Expression.random_loc.

The concrete expression module is named Pl_expression for Project
Language.