Fix simple checker; now appears to work
First bug fixed. Before, the negation of err_reached was reached,
which, when coupled with assignment of that reference to the result
of recursive calls, meant that result returned alernated between true
and false as the depth in the recursion stack varied. Fixed by using
no_err_reached instead, requiring no negation.

Second bug fixed. Before, if there were no (global) transitions
available for a particular thread, the search stopped. However, this did
not account for the fact that errors are often reached *after* the last
access of a global object in a thread. Fixed by exploring local thread
to completion when no transitions available.

Also added more debugging statements, controlled by a module-wide flag.