Function generating string for each test 34 thread
This means it is very easy to modify the number of threads used in
test 34 (this idea should be copied and used for earlier test cases),
as well as meaning that each does not spent its first few transitions
reading constant global variables, which takes a lot of model-checking
effort.

Also add a couple of examples to demonstrate how much effort such reads
take.