Implement sound stateful DPOR
Following algorithm by Yang et al. in SPIN '08, using a "visible
operation dependency graph" to conservatively update the backtrack sets
of the transitions in the currently explored sequence whenever we prune
the search because we reach a state we have already seen.

Haven't implemented the version using the light-weight state capturing
scheme yet; the whole state is used.

Seems to reduce state space as expected, but can take a lot of time to
execute.