Fix simple sleep sets
If sleep set isn't empty, the current state isn't
a deadlock as there are other transitions which
have already been explored.

Not sure sure about whether to add only enabled transitions
or not, or whether it even matters.

Should be fairly straightforward to add sleep sets to DPOR
now I have the understanding.