Vastly improve SPOR time efficiency
Use a table to store the possible locations
that each thread can access, and calculate it
only once, at the beginning, as this must
include all locations that can be accessed
later. Although it give too high a number,
this computation is expensive, so doing it
only once saves a lot of time later.

Also choose an enabled transition if possible,
rather than just going with process 0 regardless.