First draft of clock vectors section
Seems to be quite long and involved, but
I'm not sure how else to phrase it
without glossing over it, as they seem
to have in the paper. It's certainly
not obvious to me. Maybe there is a more
obvious reason that the process clock
and the object clock are updated to the
same vector? I suspect there is a
better way of phrasing it relating to this,
which could be looked at to improve
the writing and to reduce the word count.